/board/id の整形まで

+++ server/api/boards/[id].delete.ts
@@ -1,10 +1,17 @@
 import { deleteBoard } from "~~/server/utils/board"
 export default defineEventHandler(async (event) => {
+    const { user } = await requireAdminSession(event)
+
     const { id } = getRouterParams(event)
-    // if (!id) {
-    //     throw createError({ statusCode: 400, statusMessage: "Bad Request", message: "Missing board id" })
-    // }
     const numid = Number(id);
+
+    const board = await findBoardById(numid)
+    if (!board) {
+        throw createError({ statusCode: 400, statusMessage: "Bad Request", message: "Nothind data" })
+    }
+    if (board.authorId !== user.id) {
+        throw createError({ statusCode: 403, statusMessage: "Forbidden", message: "Not own" })
+    }
     const deleted = await deleteBoard(numid)
     if (!deleted) {
         throw createError({ statusCode: 404, statusMessage: "Not Found", message: "Board not found" })
